mejoras y despliegue

+++ src/main/java/com/svalero/glutenvoid/controller/UserController.java
@@ -91,13 +91,41 @@ public class UserController {
     }
 
     @DeleteMapping("/users/{id}")
-    public ResponseEntity<String> deleteUser(@PathVariable long id) throws UserNotFoundException{
-        userService.deleteUser(id);
-        String deleteMessage = "User deleted successfully";
-        logger.info("Usuario borrado exitosamenete");
-        return  ResponseEntity.status(HttpStatus.NO_CONTENT).build();
+    public ResponseEntity<String> deleteUser(@PathVariable long id) {
+        try {
+            // Obtener el nombre de usuario del contexto de seguridad
+            String username = (String) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
 
+            // Buscar el usuario actual por su nombre de usuario
+            User currentUser = userService.findByUsername(username)
+                    .orElseThrow(() -> new UserNotFoundException("User not found: " + username));
+
+            // Buscar el usuario que se desea eliminar por su ID
+            User deleteUser = userService.findById(id);
+
+            // Verificar si el usuario actual tiene permiso para eliminar la cuenta
+            if (deleteUser.getId() != currentUser.getId() && !currentUser.isAdmin()) {
+                logger.error("Usuario no autorizado para eliminar esta cuenta: " + username);
+                throw new AccessDeniedException("Usuario no autorizado para eliminar esta cuenta");
+            }
+
+            // Eliminar el usuario
+            userService.deleteUser(id);
+            String deleteMessage = "User deleted successfully";
+            logger.info("Usuario borrado exitosamente: " + deleteUser.getUsername());
+            return ResponseEntity.status(HttpStatus.NO_CONTENT).body(deleteMessage);
+        } catch (UserNotFoundException e) {
+            logger.error("Error al intentar eliminar el usuario", e);
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(e.getMessage());
+        } catch (AccessDeniedException e) {
+            logger.error("Acceso denegado al intentar eliminar el usuario", e);
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(e.getMessage());
+        } catch (Exception e) {
+            logger.error("Error inesperado al intentar eliminar el usuario", e);
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Error interno del servidor");
+        }
     }
+
 
     @PatchMapping("/users/{id}")
     public ResponseEntity<User> updateUserPartially(@PathVariable long id, @RequestBody Map<String, Object> updates)
